+++ src/hs_input.h
@@ -50,7 +50,7 @@ void hs_init_input_buffer(hs_input_buffer* b, size_t max_message_size);
 void hs_free_input_buffer(hs_input_buffer* b);
 bool hs_expand_input_buffer(hs_input_buffer* b, size_t len);
 
-int hs_open_file(hs_input* hsi, const char* subdir, size_t id);
+int hs_open_file(hs_input* hsi, const char* subdir, unsigned long long id);
 
 size_t hs_read_file(hs_input* hsi);
